move function to utils and updates objects
