+++ client/src/components/common/SearchDropDown.tsx
@@ -1,11 +1,7 @@
 import ArrowDownIcon from '@/assets/Icons/ArrowDownIcon';
 import { removeAccents } from '@/utils/Formatter';
 import { useEffect, useState } from 'react';
-
-export type Option = {
-  label: string;
-  key: string;
-};
+import { Option } from '@/types/Option';
 
 type DropdownProps = {
   id: string;
