+++ client/src/types/Option.ts
@@ -1,0 +1,4 @@
+export type Option = {
+  label: string;
+  key: string;
+};
