+++ client/src/types/Subject.ts
@@ -1,0 +1,6 @@
+export type Subject = {
+  id: number;
+  name: string;
+  code: string;
+  credits: number;
+};
